refactor: Move validate user email files to their own package
